Push context prior to consistency check
This imposes the overhead of memcpy() over the freshly copied area which
is guaranteed not to fail, however such approach has one remarkable
advantage - function about to be called is available during report
generation.

If stack area is already corrupted at that stage the report will state:
    "corruption detected PRE-CALL to foo() in parent()"
instead of:
    "corruption detected PRE-CALL to <unknown> from parent()"

In a nutshell. This change is to have the most-recent `callee` name
available during report generation, not to have that extra chunk of memory.